✨ feat[#106] : 수정된 UI 반영
