+++ src/pages/EventReview.module.css
@@ -1,17 +1,17 @@
 .content {
-    position: relative;
-    max-width: 393px;
-    width: 100%;
-    height: 100dvh;
-    background-color: var(--white);
-    margin-left: auto;
-    margin-right: auto;
-    position: relative;
-    box-sizing: border-box;
-  }
+  position: relative;
+  max-width: 393px;
+  width: 100%;
+  height: 100dvh;
+  background-color: var(--white);
+  margin-left: auto;
+  margin-right: auto;
+  position: relative;
+  box-sizing: border-box;
+}
 
 .container {
-  flex:1;
+  flex: 1;
   background-color: var(--white);
   padding: 0rem;
   height: 100vh;
@@ -24,7 +24,7 @@
 }
 
 .content::-webkit-scrollbar {
-    display: none;
+  display: none;
 }
 
 .progressBar {
@@ -39,29 +39,35 @@
 }
 
 .progressFill {
-  width: 40%;
+  width: 50%;
+  height: 0.2rem;
+  background-color: var(--primary-color);
+}
+
+.progressFillFull {
+  width: 100%;
   height: 0.2rem;
   background-color: var(--primary-color);
   /* border-radius: 2px; */
 }
 
 .footer {
-    background-color: var(--white);
-    width: 100%;
-    max-width: 393px;
-    margin: 0;
-    padding: 14px 16px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 10px;
+  background-color: var(--white);
+  width: 100%;
+  max-width: 393px;
+  margin: 0;
+  padding: 14px 16px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 10px;
 
-    position: fixed;
-    bottom: 0;
-    box-sizing: border-box;
-    border-top: 1px solid var(--color-gray20);
-    z-index: 999;
+  position: fixed;
+  bottom: 0;
+  box-sizing: border-box;
+  border-top: 1px solid var(--color-gray20);
+  z-index: 999;
 }
 
 .textContainer {
@@ -84,40 +90,40 @@
 }
 
 .confirmButton {
-    width: 100%;
-    height: 54px;
-    background-color: var(--color-gray20);
-    border-radius: 13px;
-    border: none;
+  width: 100%;
+  height: 54px;
+  background-color: var(--color-gray20);
+  border-radius: 13px;
+  border: none;
 
-    color: var(--color-gray60);
-    font-family: "Spoqa Han Sans Neo";
-    font-size: 16px;
-    font-weight: 500;
-    letter-spacing: -0.6px;
-    line-height: 22px;
+  color: var(--color-gray60);
+  font-family: "Spoqa Han Sans Neo";
+  font-size: 16px;
+  font-weight: 500;
+  letter-spacing: -0.6px;
+  line-height: 22px;
 }
 
 .confirmButton.active {
-    background-color: var(--primary-color);
-    color: var(--white);
-    cursor: pointer;
+  background-color: var(--primary-color);
+  color: var(--white);
+  cursor: pointer;
 }
 
 .optionContainer {
-    margin-top: 90px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    padding: 0 1rem;
-    gap: 40px;
-    padding-bottom: 10rem;
+  margin-top: 90px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  padding: 0 1rem;
+  gap: 40px;
+  padding-bottom: 10rem;
 }
 
 .divider {
-    margin: 0px;
-    padding:0px;
-    border: 0.5px solid var(--color-gray20);
+  margin: 0px;
+  padding: 0px;
+  border: 0.5px solid var(--color-gray20);
 }
 
 /* 주소필드 */
@@ -125,7 +131,7 @@
   width: 100%;
   height: 63px;
   padding: 15px;
-  margin:0;
+  margin: 0;
   display: flex;
   flex-direction: row;
   align-items: center;
