+++ src/recoil/event/eventReviewFormState.ts
@@ -61,9 +61,9 @@ export const eventReviewFormState = atom<EventReviewForm>({
   },
 });
 
-// 버튼 활성화 같은 "계산 상태"는 selector로
-export const eventReviewFormValidState = selector({
-  key: "eventReviewFormValidState",
+// 1단계 버튼 활성화 (대학교 ~ 방 사진 인증까지)
+export const eventReviewStep1ValidState = selector({
+  key: "eventReviewStep1ValidState",
   get: ({ get }) => {
     const f = get(eventReviewFormState);
 
@@ -81,22 +81,40 @@ export const eventReviewFormValidState = selector({
     const prosOk = f.pros.length >= 3 && f.pros.length <= 5;
     const consOk = f.cons.length >= 3 && f.cons.length <= 5;
 
-    const reviewOk = f.reviewText.trim().length >= 20;
-    // 휴대폰 번호는 하이픈 제거 후 정확히 11자리여야 함
-    const phoneNumbersOnly = f.phone.replace(/[^\d]/g, "");
-    const phoneOk = phoneNumbersOnly.length === 11;
-    const agreeOk = f.agreeMarketing && f.agreePrivacy;
-
     return (
       hasUniversity &&
       // TODO: 주소 검색 기능 활성화 시 아래 주석 해제
       // hasAddress &&
       hasPrice &&
       prosOk &&
-      consOk &&
-      reviewOk &&
-      phoneOk &&
-      agreeOk
+      consOk
     );
+  },
+});
+
+// 2단계 버튼 활성화 (찐 후기 작성 + 이벤트 참여 정보)
+export const eventReviewStep2ValidState = selector({
+  key: "eventReviewStep2ValidState",
+  get: ({ get }) => {
+    const f = get(eventReviewFormState);
+
+    const reviewOk = f.reviewText.trim().length >= 20;
+    // 휴대폰 번호는 하이픈 제거 후 정확히 11자리여야 함
+    const phoneNumbersOnly = f.phone.replace(/[^\d]/g, "");
+    const phoneOk = phoneNumbersOnly.length === 11;
+    const agreeOk = f.agreeMarketing && f.agreePrivacy;
+
+    return reviewOk && phoneOk && agreeOk;
+  },
+});
+
+// 전체 버튼 활성화 같은 "계산 상태"는 selector로
+export const eventReviewFormValidState = selector({
+  key: "eventReviewFormValidState",
+  get: ({ get }) => {
+    const step1Valid = get(eventReviewStep1ValidState);
+    const step2Valid = get(eventReviewStep2ValidState);
+
+    return step1Valid && step2Valid;
   },
 });
